# packages/wallet-desktop-openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "{{VERSION}}",
    "title": "i3M Wallet API",
    "contact": {
      "name": "Juan Hernández Serrano",
      "email": "j.hernandez@upc.edu"
    },
    "license": {
      "name": "MIT"
    },
    "description": "i3M-Wallet interface. i3M-Wallet will run as a microservice that will be called by the i3M-SDK. It is going to be developed as a desktop or mobile app.\n\nThere will be a SW implemetation of all the funcionalities and all the i3M-Wallet providers, namely the Cloud Wallet and the HW Wallet, will replace some of them.\n\ni3m-wallet should support the use of:\n- *accounts*: two kinds of accounts:\n  - identity: a public-private key pair with some contextual information\n  - secrets: cryptographic keys for symmetric crypto with some contextual information\n- *vault*: a secure storage for arbitrary objects\n"
  },
...
  "paths": {
    "/accounts": {
      "get": {
        "summary": "List all accounts or selected accounts",
        "operationId": "accountList",
        "x-eov-operation-handler": "accounts",
        "tags": [
          "accounts"
        ],
        "parameters": [
          {
            "name": "account_ids",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": false,
            "required": false,
            "allowReserved": true
          },
          {
            "name": "props",
            "in": "query",
            "description": "a comma-separated list of props to retrieve from the account. Props differ depending on the type of account ('identity' or 'secret')",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "id",
                "type",
                "name"
              ]
            },
            "style": "form",
            "explode": false,
            "required": false,
            "allowReserved": true
          }
        ],
        "responses": {
          "200": {
            "description": "A paged array of accounts. Only the props requested will be returned. Security policies may prevent some props from being returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListOutput"
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an account",
        "operationId": "accountCreate",
        "x-eov-operation-handler": "accounts",
        "tags": [
          "accounts"
        ],
        "requestBody": {
          "description": "Creation option. Whether an 'identity' or a 'secret', an optional name and comment, and addtional wallet-provider's specific options",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "the ID and type of the created account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountCreateOutput"
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/sign": {
      "post": {
        "summary": "sign the input",
        "operationId": "accountSign",
        "x-eov-operation-handler": "accounts",
        "tags": [
          "accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountSignInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An object containing the a JSON serialization of a JWS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSignOutput"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/verify": {
      "post": {
        "summary": "verify a signature",
        "operationId": "accountVerify",
        "x-eov-operation-handler": "accounts",
        "tags": [
          "accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountVerifyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An object containing whether the signature has been verified or not.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountVerifyOutput"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/encrypt": {
      "post": {
        "summary": "encrypt the input",
        "operationId": "accountEncrypt",
        "x-eov-operation-handler": "accounts",
        "tags": [
          "accounts"
        ],
        "requestBody": {
          "description": "An object containing:\n- `accountId`: [OPTIONAL] if accountId is set, it will be used; otherwise, the end user will have to interactively select one account.\n- `input`: base64url-encoded versions of the plaintext to encrypt, and an optional iv (random one will be chosen if not provided).\n",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountEncryptInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An object containing the base64url-encoded ciphertext.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEncryptOutput"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/decrypt": {
      "post": {
        "summary": "decrypt the input",
        "operationId": "accountDecrypt",
        "x-eov-operation-handler": "accounts",
        "tags": [
          "accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDecryptInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An object containing the base64url-encoded decrypted plaintext",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDecryptOutput"
                }
              }
            }
          },
          "default": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "AccountCreateInput": {
        "title": "AccountCreateInput",
        "description": "Besides the here defined options, provider specific properties should be added here if necessary, e.g. \"path\" for BIP21 wallets, or the key algorithm (if the wallet supports multiple algorithm).\n",
        "type": "object",
        "properties": {
          "type": {
            "default": "Identity"
          },
          "name": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "AccountCreateOutput": {
        "title": "AccountCreateOutput",
        "description": "It returns the account id and type\n",
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": true,
        "required": [
          "type",
          "id"
        ]
      },
      "AccountDecryptInput": {
        "title": "AccountDecryptInput",
        "description": "An object containing:\n- `accountId`: [OPTIONAL] if accountId is set, it will be used; otherwise, the end user will have to interactively select one account.\n- `ciphertext`: an object containing base64url-encoded versions of the ciphertext, and the iv used to encrypt it (if used).\n",
        "type": "object",
        "properties": {
          "accountId": {
            "$ref": "#/components/schemas/AccountId"
          },
          "ciphertext": {
            "$ref": "#/components/schemas/JWEFlattenedJSON"
          }
        },
        "required": [
          "ciphertext"
        ]
      },
      "AccountDecryptOutput": {
        "title": "AccountDecryptOutput",
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]+$",
        "description": "The decrypted message Base64Url"
      },
      "AccountEncryptInput": {
        "title": "AccountEncryptInput",
        "description": "An object containing:\n- `accountId`: [OPTIONAL] if accountId is set, it will be used; otherwise, the end user will have to interactively select one account.\n- `input`: base64url-encoded versions of the plaintext to encrypt, and an optional iv (random one will be chosen if not provided).\n",
        "type": "object",
        "properties": {
          "accountId": {
            "$ref": "#/components/schemas/AccountId"
          },
          "encryptInput": {
            "type": "object",
            "properties": {
              "plaintext": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]+$",
                "description": "the plaintext to be encrypted in Base64Url"
              },
              "iv": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]+$",
                "description": "OPTIONAL. Base64Url-encoded iv to use for encryption (if required)"
              }
            },
            "required": [
              "plaintext"
            ]
          }
        },
        "required": [
          "encryptInput"
        ]
      },
      "AccountEncryptOutput": {
        "title": "AccountEncryptOutput",
        "allOf": [
          {
            "$ref": "#/components/schemas/JWEFlattenedJSON"
          }
        ]
      },
...
      "AccountListOutput": {
        "title": "AccountListOutput",
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Account"
        }
      },
      "AccountSignInput": {
        "title": "AccountSignInput",
        "description": "An object containing:\n- `accountId`: [OPTIONAL] if accountId is set, it will be used; otherwise, the end user will have to interactively select one account.\n- `messageToSign`: It can be either an plain object or a string (preferably a BASE64URL). Examples:\n  `eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImRpZDpldGhyOjB4YjljNTcxNDA4OTQ3OGEzMjdmMDkxOTc5ODdmMTZmOWU1ZDkzNmU4YSNrZXktMSJ9`.\n",
        "type": "object",
        "properties": {
          "accountId": {
            "$ref": "#/components/schemas/AccountId"
          },
          "messageToSign": {
            "$ref": "#/components/schemas/AccountSignInputMessageToSign"
          },
          "output": {
            "type": "string",
            "enum": [
              "raw",
              "jws"
            ],
            "default": "raw"
          }
        },
        "required": [
          "messageToSign"
        ]
      },
      "AccountSignInputMessageToSign": {
        "title": "AccountSignInputMessageToSign",
        "oneOf": [
          {
            "type": "object",
            "description": "an object to sign",
            "additionalProperties": true,
            "example": {
              "msg": "hello new world",
              "date": "2050-01-01"
            }
          },
          {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "base64url-encoded binary stream",
            "example": "SGVsbG8gV29ybGQ"
          }
        ]
      },
      "AccountSignOutput": {
        "title": "AccountSignOutput",
        "allOf": [
          {
            "$ref": "#/components/schemas/Signature"
          }
        ]
      },
      "AccountVerifyInput": {
        "title": "AccountVerifyInput",
        "description": "An object containing:\n- `accountId`: [OPTIONAL] if accountId is set, it will be used; otherwise, the end user will have to interactively select one account.\n- `signature`: either a JWS JSON serialized object or a base64url-encoded binary stream\n",
        "type": "object",
        "properties": {
          "accountId": {
            "$ref": "#/components/schemas/AccountId"
          },
          "signature": {
            "$ref": "#/components/schemas/Signature"
          }
        },
        "required": [
          "signature"
        ]
      },
      "AccountVerifyOutput": {
        "title": "AccountVerifyOutput",
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean"
          }
        },
        "required": [
          "verified"
        ],
        "description": "whether the signature was properly verified (true) or not (false)."
      },
      "AccountId": {
        "title": "AccountId",
        "type": "string",
        "example": "m/44'/60'/0'/0/2"
      },
      "AccountType": {
        "title": "AccountType",
        "type": "string",
        "enum": [
          "Identity",
          "Secret"
        ]
      },
      "Account": {
        "title": "Account",
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/AccountId"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "key": {
            "$ref": "#/components/schemas/JWK"
          }
        },
        "additionalProperties": true,
        "discriminator": {
          "propertyName": "type"
        },
        "required": [
          "id"
        ]
      },
...
      "Signature": {
        "title": "Signature",
        "oneOf": [
          {
            "$ref": "#/components/schemas/JWSFlattenedJSON"
          },
          {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "base64url-encoded binary stream"
          }
        ]
      },
...
      "ApiError": {
        "type": "object",
        "title": "Error",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "jwk": {
        "type": "string",
        "description": "The \"jwk\" (JSON Web Key) when in a JWS header is the public key that corresponds to the key used to digitally sign the JWS. This key is represented as a JSON Web Key [JWK]\n",
        "pattern": "^[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$"
      },
...
      "JWK": {
        "type": "object",
        "title": "JWK",
        "properties": {
          "kty": {
            "$ref": "#/components/schemas/kty"
          },
          "use": {
            "$ref": "#/components/schemas/use"
          },
          "key_ops": {
            "$ref": "#/components/schemas/key_ops"
          },
          "alg": {
            "$ref": "#/components/schemas/alg"
          },
          "kid": {
            "$ref": "#/components/schemas/kid"
          },
          "x5u": {
            "$ref": "#/components/schemas/x5u"
          },
          "x5c": {
            "$ref": "#/components/schemas/x5c"
          },
          "x5t": {
            "$ref": "#/components/schemas/x5t"
          },
          "x5t#S256": {
            "$ref": "#/components/schemas/x5tS256"
          }
        },
        "additionalProperties": true,
        "example": {
          "kty": "EC",
          "crv": "P-256",
          "x": "f83OJ3D2xF1Bg8vub9tLe1gHMzV76e8Tus9uPHvRVEU",
          "y": "x_FEzRu9m36HLN_tue659LNpXW6pCyStikYjKIWI5a0",
          "kid": "12"
        }
      },
      "JWEFlattenedJSON": {
        "title": "JWEFlattenedJSON",
        "type": "object",
        "description": "The flattened JWE JSON Serialization syntax is based upon the general syntax, but flattens it, optimizing it for the single-recipient case. It flattens it by removing the \"recipients\" member and instead placing those members defined for use in the \"recipients\" array (the \"header\" and \"encrypted_key\" members) in the top-level JSON object (at the same level as the \"ciphertext\" member).\n\nThe \"recipients\" member MUST NOT be present when using this syntax. Other than this syntax difference, JWE JSON Serialization objects using the flattened syntax are processed identically to those using the general syntax.\n\nIn summary, the syntax of a JWE using the flattened JWE JSON\n\nSerialization is as follows:\n  {\n    \"protected\":\"<integrity-protected header contents>\",\n    \"unprotected\":<non-integrity-protected header contents>,\n    \"header\":<more non-integrity-protected header contents>,\n    \"encrypted_key\":\"<encrypted key contents>\",\n    \"aad\":\"<additional authenticated data contents>\",\n    \"iv\":\"<initialization vector contents>\",\n    \"ciphertext\":\"<ciphertext contents>\",\n    \"tag\":\"<authentication tag contents>\"\n  }\n\nNote that when using the flattened syntax, just as when using the\ngeneral syntax, any unprotected Header Parameter values can reside in\neither the \"unprotected\" member or the \"header\" member, or in both.\n",
        "example": {
          "protected": "eyJlbmMiOiJBMTI4Q0JDLUhTMjU2In0",
          "unprotected": {
            "jku": "https://server.example.com/keys.jwks"
          },
          "header": {
            "alg": "A128KW",
            "kid": "7"
          },
          "encrypted_key": "6KB707dM9YTIgHtLvtgWQ8mKwboJW3of9locizkDTHzBC2IlrT1oOQ\"",
          "iv": "AxY8DCtDaGlsbGljb3RoZQ",
          "ciphertext": "KDlTtXchhZTGufMYmOYGS4HffxPSUrfmqCHXaI9wOGY\"",
          "tag": "Mz-VPPyU4RlcuYv1IwIvzw"
        },
        "properties": {
          "protected": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "BASE64URL(UTF8(JWE Protected Header))\n\nThe JWE Protected Header declares:\n  - `alg`: algorithm to encrypt a fresh and randomly generated Content Encryption Key (CEK) `encrypted_key`\n  - `enc`: the content encryption algorithm. It should be a symmetric Authenticated Encryption with Associated Data (AEAD) algorithm. Contents will we encrypted using `enc` algorithm with the `encrypted_key`\n  - `kid`: an identifier of the key that shold be used to decrypt the `encrypted_key`\n\nThe \"protected\" member MUST be present and contain the value BASE64URL(UTF8(JWE Protected Header)) when the JWE Protected Header value is non-empty; otherwise, it MUST be absent. These Header Parameter values are integrity protected.\n\nAn example JWE Protected header is:\n`{\"alg\":\"ES256\",\"enc\":\"A256GCM\",\"kid\":\"did:ethr:0xb9c5714089478a327f09197987f16f9e5d936e8a#key-1\"}`\n",
            "example": "eyJlbmMiOiJBMjU2R0NNIiwiYWxnIjoiRVMyNTYiLCJraWQiOiJkaWQ6ZXRocjoweGI5YzU3MTQwODk0NzhhMzI3ZjA5MTk3OTg3ZjE2ZjllNWQ5MzZlOGEja2V5LTEifQ"
          },
          "unprotected": {
            "type": "object",
            "additionalProperties": true,
            "description": "JWE Unprotected Header\n\nNon-integrity-protected header contents\n\nThe \"unprotected\" member MUST be present and contain the value JWE Shared Unprotected Header when the JWE Shared Unprotected Header value is non-empty; otherwise, it MUST be absent. This value is represented as an unencoded JSON object, rather than as a string. These Header Parameter values are not integrity protected.\n"
          },
          "header": {
            "type": "object",
            "properties": {
              "alg": {
                "$ref": "#/components/schemas/alg"
              },
              "enc": {
                "$ref": "#/components/schemas/enc"
              },
              "zip": {
                "$ref": "#/components/schemas/zip"
              },
              "jku": {
                "$ref": "#/components/schemas/jku"
              },
              "jwk": {
                "$ref": "#/components/schemas/jwk"
              },
              "kid": {
                "$ref": "#/components/schemas/kid"
              },
              "x5u": {
                "$ref": "#/components/schemas/x5u"
              },
              "x5c": {
                "$ref": "#/components/schemas/x5c"
              },
              "x5t": {
                "$ref": "#/components/schemas/x5t"
              },
              "x5t#S256": {
                "$ref": "#/components/schemas/x5tS256"
              },
              "typ": {
                "$ref": "#/components/schemas/typ"
              },
              "cty": {
                "$ref": "#/components/schemas/cty"
              },
              "crit": {
                "$ref": "#/components/schemas/crit"
              }
            },
            "additionalProperties": true,
            "description": "More non-integrity-protected header contents\n\nThe \"header\" member MUST be present and contain the value JWE Per-Recipient Unprotected Header when the JWE Per-Recipient Unprotected Header value is non-empty; otherwise, it MUST be absent. This value is represented as an unencoded JSON object, rather than as a string. These Header Parameter values are not integrity protected.\n"
          },
          "encrypted_key": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "BASE64URL(JWE Encrypted Key)\nIt should be decrypted with the key identified in the `kid` protected header parameter. Once decrypted, it can be used to decrypt the `ciphertext`\n"
          },
          "aad": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "OPTIONAL. BASE64URL(JWE additional Authenticated Data contents). A JWE AAD value can be included to supply a base64url-encoded value to be integrity protected but not encrypted.\n"
          },
          "iv": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "OPTIONAL. BASE64URL(JWE Initialization Vector)\nInitialization Vector value used when encrypting the plaintext\n"
          },
          "ciphertext": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "BASE64URL(JWE Ciphertext)"
          },
          "tag": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "OPTIONAL. BASE64URL(JWE Authentication Tag)\nAn output of an AEAD operation that ensures the integrity of the ciphertext and the Additional Authenticated Data.  Note that some algorithms may not use an Authentication Tag, in which case this value is the empty octet sequence.\n"
          }
        }
      },
      "JWSFlattenedJSON": {
        "title": "JWSFlattenedJSON",
        "description": "A JWS with flattened JSON serialization. Optional unprotected header properties are not implemented",
        "type": "object",
        "properties": {
          "protected": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "Base64Url encoded of the JWS Protected Headers: BASE64URL(UTF8(JWS Protected Header)).\n\nThe JWS Protected Header declares that the encoded object is a JSON Web Token (`\"typ\"=\"JWT\"`), the signature algorithm `alg`, and and identifier of the verification key (`kid`)\n            \nExamples values for `alg` could be:\n  - `HS256`: HMAC using SHA-256 hash algorithm\n  - `HS384`: HMAC using SHA-384 hash algorithm\n  - `HS512`: HMAC using SHA-512 hash algorithm\n  - `ES256`: ECDSA using P-256 curve and SHA-256 hash algorithm\n  - `ES384`: ECDSA using P-384 curve and SHA-384 hash algorithm\n  - `ES512`: ECDSA using P-521 curve and SHA-512 hash algorithm\n\nAn example JWS Protected header is:\n`{\"typ\":\"JWT\",\"alg\":\"ES256\",\"kid\":\"did:ethr:0xb9c5714089478a327f09197987f16f9e5d936e8a#key-1\"}`\n",
            "example": "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImRpZDpldGhyOjB4YjljNTcxNDA4OTQ3OGEzMjdmMDkxOTc5ODdmMTZmOWU1ZDkzNmU4YSNrZXktMSJ9"
          },
          "header": {
            "description": "For a JWS, the members of the JSON object(s) representing the JOSE Header describe the digital signature or MAC applied to the JWS Protected Header and the JWS Payload and optionally additional properties of the JWS. The Header Parameter names within the JOSE Header MUST be unique; JWS parsers MUST either reject JWSs with duplicate Header Parameter names or use a JSON parser that returns only the lexically last duplicate member name, as specified in Section 15.12 (\"The JSON Object\") of ECMAScript 5.1 [ECMAScript].\n\nImplementations are required to understand the specific Header Parameters defined by this specification that are designated as \"MUST be understood\" and process them in the manner defined in this specification.  All other Header Parameters defined by this specification that are not so designated MUST be ignored when not understood.  Unless listed as a critical Header Parameter, per   Section 4.1.11, all Header Parameters not defined by this specification MUST be ignored when not understood.\n",
            "type": "object",
            "properties": {
              "alg": {
                "$ref": "#/components/schemas/alg"
              },
              "jku": {
                "$ref": "#/components/schemas/jku"
              },
              "jwk": {
                "$ref": "#/components/schemas/jwk"
              },
              "kid": {
                "$ref": "#/components/schemas/kid"
              },
              "x5u": {
                "$ref": "#/components/schemas/x5u"
              },
              "x5c": {
                "$ref": "#/components/schemas/x5c"
              },
              "x5t": {
                "$ref": "#/components/schemas/x5t"
              },
              "x5t#S256": {
                "$ref": "#/components/schemas/x5tS256"
              },
              "typ": {
                "$ref": "#/components/schemas/typ"
              },
              "cty": {
                "$ref": "#/components/schemas/cty"
              },
              "crit": {
                "$ref": "#/components/schemas/crit"
              }
            },
            "additionalProperties": true,
            "example": {
              "kid": "e9bc097a-ce51-4036-9562-d2ade882db0d"
            }
          },
          "payload": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "Base64Url encoded of the JWS Pyload: BASE64URL(JWS Payload).\nThe payload is the binary array that it has been signed. It is often a UTF-8 representation of a JSON object that includes the `signed` object along with some other props assigned by the signer.\n\nAn example payload could be:\n`{\"iss\":\"did:ethr:0xb9c5714089478a327f09197987f16f9e5d936e8a\",\"iat\":1611076613,\"signed\":\"aTMtbWFya2V0IGlzIHdoYXQgaXQgaXMuIERvIHlvdSBrbm93IHdoYXQgaXMgaXQ_\"}`\n",
            "example": "eyJpc3MiOiJkaWQ6ZXRocjoweGI5YzU3MTQwODk0NzhhMzI3ZjA5MTk3OTg3ZjE2ZjllNWQ5MzZlOGEiLCJpYXQiOjE2MTEwNzY2MTMsInNpZ25lZCI6ImFUTXRiV0Z5YTJWMElHbHpJSGRvWVhRZ2FYUWdhWE11SUVSdklIbHZkU0JyYm05M0lIZG9ZWFFnYVhNZ2FYUV8ifQ"
          },
          "signature": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "The signature or HMAC of the JWS Signing Input ASCII(BASE64URL(UTF8(JWS Protected Header)) || '.' || BASE64URL(JWS Payload)) with the algoritm `alg`\n",
            "example": "LSOC5nsxpqs3XeRh_uhBbGlF1uqyOxFNmHfUua66Fw0iPM1plHh01V4nVrz10Hq8_6oOowtU9ePKvxPTDYZo4g"
          }
        },
        "required": [
          "protected",
          "payload",
          "signature"
        ],
        "example": {
          "payload": "eyJpc3MiOiJqb2UiLA0KICJleHAiOjEzMDA4MTkzODAsDQogImh0dHA6Ly9leGFtcGxlLmNvbS9pc19yb290Ijp0cnVlfQ",
          "protected": "eyJhbGciOiJFUzI1NiJ9",
          "header": {
            "kid": "e9bc097a-ce51-4036-9562-d2ade882db0d"
          },
          "signature": "DtEhU3ljbEg8L38VWAfUAqOyKAM6-Xx-F4GawxaepmXFCgfTjDxw5djxLa8ISlSApmWQxfKTUJqPP3-Kg6NU1Q"
        }
      },
...
    }
  }
}